Add files for JS functions
Add worker_main, which handles all of the orginazational
aspects of the work to be done.  This handles creating all
of the workers, and dealing with the promises, which is
being called with Promise.all, so that we are only generating our
approximations once we have all of the results from each of our
workers/promises.

Add worker which is the base of the Pi approximation.  Here the
Monte Carlo approximation is implemented.